requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Base Dependencies
# Packages required by MOST or ALL nodes (Worker, CNC, Observability).
# ─────────────────────────────────────────────────────────────────────────────

# --- Core & Cloud ---
boto3
botocore
psutil
pyyaml
python-dotenv

# --- Orchestration & AI ---
temporalio
litellm
google-genai
qdrant-client
pydantic
opik>=1.0.0

# --- Utilities ---
httpx
redis
requests
beautifulsoup4
prometheus-client>=0.17.0

# --- Auth & Security ---
PyJWT
bcrypt
psycopg2-binary

# --- Development & Linting (can be moved to a dev-requirements.txt) ---
flake8
pytest
pytest-asyncio
bandit
ansible-lint
fastapi
uvicorn
sse-starlette

# --- Tool plugin system ---
asyncpg           # async Postgres driver for tool config DB
psycopg2-binary   # sync Postgres driver for config_db
cryptography      # AES-256-GCM credential encryption

# --- Scheduler ---
croniter          # Cron expression parsing
playwright>=1.40.0

# --- Document Processing ---
pypdf>=4.0        # PDF text extraction
python-docx>=1.1  # DOCX parsing
